Modularize general PDF Outfile functions from wkhtmltopdf-specific functions
