+++ internal/bagoup/export_test.go
@@ -234,7 +234,7 @@ func TestExportChats(t *testing.T) {
 					dbMock.EXPECT().GetMessageIDs(1),
 					osMock.EXPECT().MkdirAll("messages-export/testdisplayname", os.ModePerm),
 					osMock.EXPECT().Create("messages-export/testdisplayname/testguid.pdf").Return(chatFile, nil),
-					osMock.EXPECT().NewPDFOutFile(chatFile, gomock.Any(), false).Return(ofMocks[0]),
+					osMock.EXPECT().NewWkhtmltopdfOutFile(chatFile, gomock.Any(), false).Return(ofMocks[0]),
 					ofMocks[0].EXPECT().Stage(),
 					osMock.EXPECT().GetOpenFilesLimit().Return(256, nil),
 					ofMocks[0].EXPECT().Flush(),
@@ -264,7 +264,7 @@ func TestExportChats(t *testing.T) {
 					dbMock.EXPECT().GetMessageIDs(1),
 					osMock.EXPECT().MkdirAll("messages-export/testdisplayname", os.ModePerm),
 					osMock.EXPECT().Create("messages-export/testdisplayname/testguid.pdf").Return(chatFile, nil),
-					osMock.EXPECT().NewPDFOutFile(chatFile, gomock.Any(), false).Return(ofMocks[0]),
+					osMock.EXPECT().NewWkhtmltopdfOutFile(chatFile, gomock.Any(), false).Return(ofMocks[0]),
 					ofMocks[0].EXPECT().Stage(),
 					osMock.EXPECT().GetOpenFilesLimit().Return(256, nil),
 					ofMocks[0].EXPECT().Flush(),
