+++ opsys/mock_opsys/mock_opsys.go
@@ -243,20 +243,6 @@ func (mr *MockOSMockRecorder) Name() *gomock.Call {
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "Name", reflect.TypeOf((*MockOS)(nil).Name))
 }
 
-// NewPDFOutFile mocks base method.
-func (m *MockOS) NewPDFOutFile(arg0 afero.File, arg1 pdfgen.PDFGenerator, arg2 bool) opsys.OutFile {
-	m.ctrl.T.Helper()
-	ret := m.ctrl.Call(m, "NewPDFOutFile", arg0, arg1, arg2)
-	ret0, _ := ret[0].(opsys.OutFile)
-	return ret0
-}
-
-// NewPDFOutFile indicates an expected call of NewPDFOutFile.
-func (mr *MockOSMockRecorder) NewPDFOutFile(arg0, arg1, arg2 interface{}) *gomock.Call {
-	mr.mock.ctrl.T.Helper()
-	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "NewPDFOutFile", reflect.TypeOf((*MockOS)(nil).NewPDFOutFile), arg0, arg1, arg2)
-}
-
 // NewTxtOutFile mocks base method.
 func (m *MockOS) NewTxtOutFile(arg0 afero.File) opsys.OutFile {
 	m.ctrl.T.Helper()
@@ -269,6 +255,20 @@ func (m *MockOS) NewTxtOutFile(arg0 afero.File) opsys.OutFile {
 func (mr *MockOSMockRecorder) NewTxtOutFile(arg0 interface{}) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "NewTxtOutFile", reflect.TypeOf((*MockOS)(nil).NewTxtOutFile), arg0)
+}
+
+// NewWkhtmltopdfOutFile mocks base method.
+func (m *MockOS) NewWkhtmltopdfOutFile(arg0 afero.File, arg1 pdfgen.PDFGenerator, arg2 bool) opsys.OutFile {
+	m.ctrl.T.Helper()
+	ret := m.ctrl.Call(m, "NewWkhtmltopdfOutFile", arg0, arg1, arg2)
+	ret0, _ := ret[0].(opsys.OutFile)
+	return ret0
+}
+
+// NewWkhtmltopdfOutFile indicates an expected call of NewWkhtmltopdfOutFile.
+func (mr *MockOSMockRecorder) NewWkhtmltopdfOutFile(arg0, arg1, arg2 interface{}) *gomock.Call {
+	mr.mock.ctrl.T.Helper()
+	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "NewWkhtmltopdfOutFile", reflect.TypeOf((*MockOS)(nil).NewWkhtmltopdfOutFile), arg0, arg1, arg2)
 }
 
 // Open mocks base method.
