+++ opsys/opsys.go
@@ -65,7 +65,7 @@ type (
 		// path is returned.
 		HEIC2JPG(src string) (string, error)
 		NewTxtOutFile(afero.File) OutFile
-		NewPDFOutFile(chatFile afero.File, pdfg pdfgen.PDFGenerator, includePPA bool) OutFile
+		NewWkhtmltopdfOutFile(chatFile afero.File, pdfg pdfgen.PDFGenerator, includePPA bool) OutFile
 	}
 
 	opSys struct {
